+++ src/canvas/CanvasMeshRenderer.js
@@ -1,4 +1,4 @@
-import { default as Mesh } from '../Mesh';
+import Mesh from '../Mesh';
 
 /**
  * Renderer dedicated to meshes.
@@ -47,6 +47,7 @@ class MeshSpriteRenderer {
       );
     }
 
+    renderer.context.globalAlpha = mesh.worldAlpha;
     renderer.setBlendMode(mesh.blendMode);
 
     if (mesh.drawMode === Mesh.DRAW_MODES.TRIANGLE_MESH) {
@@ -151,9 +152,11 @@ class MeshSpriteRenderer {
     let y1 = vertices[index1 + 1];
     let y2 = vertices[index2 + 1];
 
-    if (mesh.canvasPadding > 0) {
-      const paddingX = mesh.canvasPadding / Math.abs(mesh.worldTransform.a);
-      const paddingY = mesh.canvasPadding / Math.abs(mesh.worldTransform.d);
+    const canvasPadding = mesh.canvasPadding / this.renderer.resolution;
+
+    if (canvasPadding > 0) {
+      const paddingX = canvasPadding / Math.abs(mesh.worldTransform.a);
+      const paddingY = canvasPadding / Math.abs(mesh.worldTransform.d);
       const centerX = (x0 + x1 + x2) / 3;
       const centerY = (y0 + y1 + y2) / 3;
 
@@ -226,6 +229,7 @@ class MeshSpriteRenderer {
     }
 
     context.restore();
+    this.renderer.invalidateBlendMode();
   }
 
   /**