+++ src/index.js
@@ -14,9 +14,9 @@
  *
  * @namespace Tiny.mesh
  */
-export {default as Mesh} from './Mesh';
-export {default as MeshRenderer} from './webgl/MeshRenderer';
-export {default as CanvasMeshRenderer} from './canvas/CanvasMeshRenderer';
-export {default as Plane} from './Plane';
-export {default as NineSlicePlane} from './NineSlicePlane';
-export {default as Rope} from './Rope';
+export { default as Mesh } from './Mesh';
+export { default as MeshRenderer } from './webgl/MeshRenderer';
+export { default as CanvasMeshRenderer } from './canvas/CanvasMeshRenderer';
+export { default as Plane } from './Plane';
+export { default as NineSlicePlane } from './NineSlicePlane';
+export { default as Rope } from './Rope';